RE-json file
The reason I deleted the json.js file was because the code would run
and say that var json1 is missing I simply fixed this by adding the
line of code so the html file can call it. Just a small mistake and I
prefer 2 files to keep it a little more organized.